Cast odds and maxbets back into int arrays

+++ neofoodclub/nfc.py
@@ -800,8 +800,8 @@ class NeoFoodClub:
         self._data_dict["std"] = _stds
         self._data_dict["ers"] = _ers
         self._data_dict["bins"] = _bins
-        self._data_dict["odds"] = _odds
-        self._data_dict["maxbets"] = _maxbets
+        self._data_dict["odds"] = _odds.astype(int)
+        self._data_dict["maxbets"] = _maxbets.astype(int)
 
         self._cache_bet_amount_dicts()
